# Remove global logger