--- a/server/opts.go
+++ b/server/opts.go
@@ -736,15 +736,21 @@ func RemoveSelfReference(clusterPort int, routes []*url.URL) ([]*url.URL, error)
 	var cleanRoutes []*url.URL
 	cport := strconv.Itoa(clusterPort)
 
-	selfIPs := getInterfaceIPs()
+	selfIPs, err := getInterfaceIPs()
+	if err != nil {
+		return nil, err
+	}
 	for _, r := range routes {
 		host, port, err := net.SplitHostPort(r.Host)
 		if err != nil {
 			return nil, err
 		}
 
-		if cport == port && isIPInList(selfIPs, getURLIP(host)) {
-			Noticef("Self referencing IP found: ", r)
+		ipList, err := getURLIP(host)
+		if err != nil {
+			return nil, err
+		}
+		if cport == port && isIPInList(selfIPs, ipList) {
 			continue
 		}
 		cleanRoutes = append(cleanRoutes, r)
@@ -764,19 +770,18 @@ func isIPInList(list1 []net.IP, list2 []net.IP) bool {
 	return false
 }
 
-func getURLIP(ipStr string) []net.IP {
+func getURLIP(ipStr string) ([]net.IP, error) {
 	ipList := []net.IP{}
 
 	ip := net.ParseIP(ipStr)
 	if ip != nil {
 		ipList = append(ipList, ip)
-		return ipList
+		return ipList, nil
 	}
 
 	hostAddr, err := net.LookupHost(ipStr)
 	if err != nil {
-		Errorf("Error looking up host with route hostname: %v", err)
-		return ipList
+		return nil, fmt.Errorf("Error looking up host with route hostname: %v", err)
 	}
 	for _, addr := range hostAddr {
 		ip = net.ParseIP(addr)
@@ -784,16 +789,15 @@ func getURLIP(ipStr string) []net.IP {
 			ipList = append(ipList, ip)
 		}
 	}
-	return ipList
+	return ipList, nil
 }
 
-func getInterfaceIPs() []net.IP {
+func getInterfaceIPs() ([]net.IP, error) {
 	var localIPs []net.IP
 
 	interfaceAddr, err := net.InterfaceAddrs()
 	if err != nil {
-		Errorf("Error getting self referencing address: %v", err)
-		return localIPs
+		return nil, fmt.Errorf("Error getting self referencing address: %v", err)
 	}
 
 	for i := 0; i < len(interfaceAddr); i++ {
@@ -801,10 +805,10 @@ func getInterfaceIPs() []net.IP {
 		if net.ParseIP(interfaceIP.String()) != nil {
 			localIPs = append(localIPs, interfaceIP)
 		} else {
-			Errorf("Error parsing self referencing address: %v", err)
+			return nil, fmt.Errorf("Error parsing self referencing address: %v", err)
 		}
 	}
-	return localIPs
+	return localIPs, nil
 }
 
 func processOptions(opts *Options) {