@tylertreat 2017-06-01
Results from running benchmarks before and after logging changes:
```
benchmark                         old ns/op     new ns/op     delta
Benchmark_____Pub0b_Payload-4     102           103           +0.98%
Benchmark_____Pub8b_Payload-4     108           102           -5.56%
Benchmark____Pub32b_Payload-4     160           123           -23.12%
Benchmark___Pub128B_Payload-4     205           127           -38.05%
Benchmark___Pub256B_Payload-4     223           200           -10.31%
Benchmark_____Pub1K_Payload-4     529           484           -8.51%
Benchmark_____Pub4K_Payload-4     1992          1961          -1.56%
Benchmark_____Pub8K_Payload-4     4639          4225          -8.92%
Benchmark_AuthPub0b_Payload-4     231           221           -4.33%
Benchmark____________PubSub-4     273           236           -13.55%
Benchmark____PubSubTwoConns-4     239           239           +0.00%
Benchmark____PubTwoQueueSub-4     317           287           -9.46%
Benchmark___PubFourQueueSub-4     300           255           -15.00%
Benchmark__PubEightQueueSub-4     303           270           -10.89%
Benchmark___RoutedPubSub_0b-4     761           607           -20.24%
Benchmark___RoutedPubSub_1K-4     2180          2385          +9.40%
Benchmark_RoutedPubSub_100K-4     190695        165742        -13.09%

benchmark                         old MB/s     new MB/s     speedup
Benchmark_____Pub0b_Payload-4     107.45       106.01       0.99x
Benchmark_____Pub8b_Payload-4     175.84       185.70       1.06x
Benchmark____Pub32b_Payload-4     274.54       357.46       1.30x
Benchmark___Pub128B_Payload-4     686.76       1106.86      1.61x
Benchmark___Pub256B_Payload-4     1201.65      1344.83      1.12x
Benchmark_____Pub1K_Payload-4     1962.03      2141.75      1.09x
Benchmark_____Pub4K_Payload-4     2062.49      2095.26      1.02x
Benchmark_____Pub8K_Payload-4     1768.77      1942.21      1.10x
Benchmark_AuthPub0b_Payload-4     47.54        49.59        1.04x
```