@ColinSullivan1 2017-06-02
If you add nil checks, what is the performance impact?