@derekcollison 2017-06-02
LGTM, we can add nil checks as needed in real life scenarios.